+++ src/app/app.component.ts
@@ -4,7 +4,6 @@ import {PrimeNGConfig} from 'primeng/api';
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit{