+++ src/app/demo/view/floatlabeldemo.component.ts
@@ -3,7 +3,6 @@ import {CountryService} from '../service/countryservice';
 
 @Component({
     templateUrl: './floatlabeldemo.component.html',
-    styleUrls: ['./floatlabeldemo.scss']
 })
 export class FloatLabelDemoComponent implements OnInit {
 
@@ -38,6 +37,7 @@ export class FloatLabelDemoComponent implements OnInit {
     value12: any;
 
     constructor(private countryService: CountryService) {
+
         this.cities = [
             {name: 'New York', code: 'NY'},
             {name: 'Rome', code: 'RM'},