+++ src/app/demo/view/formlayoutdemo.component.ts
@@ -14,4 +14,19 @@ export class FormLayoutDemoComponent {
         {name: 'Ohio', code: 'Ohio'},
         {name: 'Washington', code: 'Washington'}
     ];
+
+    dropdownItems = [
+        { name: 'Option 1', code: 'Option 1' },
+        { name: 'Option 2', code: 'Option 2' },
+        { name: 'Option 3', code: 'Option 3' }
+    ];
+
+    cities1: any[] = [];
+
+    cities2: any[] = [];
+
+    city1: any = null;
+
+    city2: any = null;
+
 }