+++ src/app/demo/view/inputdemo.component.ts
@@ -18,6 +18,7 @@ import {SelectItem} from 'primeng/api';
 
 	}
 
+
 	:host ::ng-deep .multiselect-custom .country-item.country-item-value {
 		padding: .25rem .5rem;
 		border-radius: 3px;
@@ -84,8 +85,6 @@ export class InputDemoComponent implements OnInit{
     valSelect2: string;
 
     valueKnob = 20;
-
-    valCheckbox: string;
 
     constructor(private countryService: CountryService) {}