+++ src/app/demo/view/invalidstatedemo.component.ts
@@ -31,7 +31,10 @@ export class InvalidStateDemoComponent implements OnInit {
 
     value9: any;
 
+    value10: any;
+
     constructor(private countryService: CountryService) {
+
         this.cities = [
             {name: 'New York', code: 'NY'},
             {name: 'Rome', code: 'RM'},