+++ src/app/demo/view/listdemo.component.ts
@@ -5,7 +5,8 @@ import {ProductService} from '../service/productservice';
 
 @Component({
     templateUrl: './listdemo.component.html',
-    styleUrls: ['./listdemo.scss']
+    styleUrls: ['../../../assets/demo/badges.scss']
+
 })
 export class ListDemoComponent implements OnInit {