+++ src/app/demo/view/mediademo.component.ts
@@ -5,14 +5,14 @@ import { Product } from '../domain/product';
 
 @Component({
     templateUrl: './mediademo.component.html',
-    styleUrls: ['./mediademo.scss'],
+    styleUrls: ['../../../assets/demo/badges.scss']
 })
 export class MediaDemoComponent implements OnInit{
 
     products: Product[];
 
     images: any[];
-
+  
     galleriaResponsiveOptions: any[] = [
         {
             breakpoint: '1024px',
@@ -31,7 +31,7 @@ export class MediaDemoComponent implements OnInit{
             numVisible: 1
         }
     ];
-
+  
     carouselResponsiveOptions: any[] = [
         {
             breakpoint: '1024px',
@@ -49,14 +49,14 @@ export class MediaDemoComponent implements OnInit{
             numScroll: 1
         }
     ];
-
-    constructor(private productService: ProductService, private photoService: PhotoService) {}
+  
+    constructor(private productService: ProductService, private photoService: PhotoService) {}  
 
     ngOnInit() {
         this.productService.getProductsSmall().then(products => {
             this.products = products;
         });
-
+  
         this.photoService.getImages().then(images => {
             this.images = images;
         });