+++ src/app/demo/view/menus/menus.component.ts
@@ -1,12 +1,10 @@
-import {Component, OnInit, ViewEncapsulation} from '@angular/core';
-import {MegaMenuItem, MenuItem} from 'primeng/api';
-
+import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+import { MegaMenuItem, MenuItem } from 'primeng/api';
 @Component({
-    templateUrl: './menusdemo.component.html',
-    styles:['.p-slidemenu { overflow-x:hidden;}'],
+    templateUrl: './menus.component.html',
     encapsulation: ViewEncapsulation.None
 })
-export class MenusDemoComponent implements OnInit {
+export class MenusComponent implements OnInit {
 
     breadcrumbItems: MenuItem[];
 
@@ -14,7 +12,7 @@ export class MenusDemoComponent implements OnInit {
 
     items: MenuItem[];
 
-    tabMenuItems: MenuItem[];
+    routeItems: MenuItem[];
 
     megaMenuItems: MegaMenuItem[];
 
@@ -25,6 +23,10 @@ export class MenusDemoComponent implements OnInit {
     slideItems: MenuItem[];
 
     menuItems: MenuItem[];
+
+    plainMenuItems: MenuItem[];
+
+    pageIndex: number = 0;
 
     ngOnInit() {
 
@@ -173,7 +175,7 @@ export class MenusDemoComponent implements OnInit {
                 separator: true
             },
             {
-                label: 'Quit', icon: 'pi pi-fw pi-sign-out'
+                label: 'Home', icon: 'pi pi-fw pi-home'
             },
         ];
 
@@ -243,6 +245,36 @@ export class MenusDemoComponent implements OnInit {
             }
         ];
 
+        this.plainMenuItems = [
+            {
+                label: 'Customers',
+                items: [
+                    {
+                        label: 'New',
+                        icon: 'pi pi-fw pi-plus'
+                    },
+                    {
+                        label: 'Edit',
+                        icon: 'pi pi-fw pi-user-edit'
+                    }
+                ]
+            },
+            {
+                label: 'Orders',
+                items: [
+                    {
+                        label: 'View',
+                        icon: 'pi pi-fw pi-list'
+                    },
+                    {
+                        label: 'Search',
+                        icon: 'pi pi-fw pi-search'
+                    }
+
+                ]
+            }
+        ];
+
         this.breadcrumbItems = [];
         this.breadcrumbItems.push({ label: 'Electronics' });
         this.breadcrumbItems.push({ label: 'Computer' });
@@ -251,12 +283,11 @@ export class MenusDemoComponent implements OnInit {
         this.breadcrumbItems.push({ label: 'Backpacks' });
         this.breadcrumbItems.push({ label: 'Item' });
 
-        this.tabMenuItems = [
-            {label: 'Overview', icon: 'pi pi-fw pi-home'},
-            {label: 'Members', icon: 'pi pi-fw pi-users'},
-            {label: 'Sales', icon: 'pi pi-fw pi-shopping-cart'},
-            {label: 'Profile', icon: 'pi pi-fw pi-user'},
-            {label: 'Settings', icon: 'pi pi-fw pi-cog'}
+        this.routeItems = [
+            {label: 'Personal', routerLink:'personal'},
+            {label: 'Seat', routerLink:'seat'},
+            {label: 'Payment', routerLink:'payment'},
+            {label: 'Confirmation', routerLink:'confirmation'},
         ];
 
         this.megaMenuItems = [
@@ -434,21 +465,6 @@ export class MenusDemoComponent implements OnInit {
                         icon: 'pi pi-fw pi-file'
                     }
                 ]
-            }
-        ];
-
-        this.stepsItems = [
-            {
-                label: 'Personal'
-            },
-            {
-                label: 'Seat'
-            },
-            {
-                label: 'Payment'
-            },
-            {
-                label: 'Confirmation'
             }
         ];
     }