+++ src/app/pages/app.crud.component.ts
@@ -1,35 +1,20 @@
 import {Component, OnInit} from '@angular/core';
-import { Product } from '../demo/domain/product';
-import { ProductService } from '../demo/service/productservice';
-import { ConfirmationService } from 'primeng/api';
-import { MessageService } from 'primeng/api';
+import {Product} from '../demo/domain/product';
+import {ProductService} from '../demo/service/productservice';
+import {ConfirmationService, MessageService} from 'primeng/api';
 
 @Component({
     templateUrl: './app.crud.component.html',
-    styleUrls: ['../demo/view/tabledemo.scss'],
-    styles: [`
-        :host ::ng-deep .p-dialog .product-image {
-            width: 150px;
-            margin: 0 auto 2rem auto;
-            display: block;
-        }
-
-        @media screen and (max-width: 960px) {
-            :host ::ng-deep .p-datatable.p-datatable-customers .p-datatable-tbody > tr > td:last-child {
-                text-align: center;
-            }
-
-            :host ::ng-deep .p-datatable.p-datatable-customers .p-datatable-tbody > tr > td:nth-child(6) {
-                display: flex;
-            }
-        }
-
-    `],
-    providers: [MessageService, ConfirmationService]
+    providers: [MessageService, ConfirmationService],
+    styleUrls: ['../../assets/demo/badges.scss']
 })
 export class AppCrudComponent implements OnInit {
 
     productDialog: boolean;
+
+    deleteProductDialog: boolean = false;
+
+    deleteProductsDialog: boolean = false;
 
     products: Product[];
 
@@ -41,18 +26,27 @@ export class AppCrudComponent implements OnInit {
 
     cols: any[];
 
-    constructor(private productService: ProductService, private messageService: MessageService,
-                private confirmationService: ConfirmationService) {}
+    statuses: any[];
 
+    rowsPerPageOptions = [5, 10, 20];
+
+    constructor(private productService: ProductService, private messageService: MessageService, private confirmationService: ConfirmationService) {}
+    
     ngOnInit() {
         this.productService.getProducts().then(data => this.products = data);
 
         this.cols = [
-            { field: 'name', header: 'Name' },
-            { field: 'price', header: 'Price' },
-            { field: 'category', header: 'Category' },
-            { field: 'rating', header: 'Reviews' },
-            { field: 'inventoryStatus', header: 'Status' }
+            {field: 'name', header: 'Name'},
+            {field: 'price', header: 'Price'},
+            {field: 'category', header: 'Category'},
+            {field: 'rating', header: 'Reviews'},
+            {field: 'inventoryStatus', header: 'Status'}
+        ];
+
+        this.statuses = [
+            {label: 'INSTOCK', value: 'instock'},
+            {label: 'LOWSTOCK', value: 'lowstock'},
+            {label: 'OUTOFSTOCK', value: 'outofstock'}
         ];
     }
 
@@ -63,16 +57,7 @@ export class AppCrudComponent implements OnInit {
     }
 
     deleteSelectedProducts() {
-        this.confirmationService.confirm({
-            message: 'Are you sure you want to delete the selected products?',
-            header: 'Confirm',
-            icon: 'pi pi-exclamation-triangle',
-            accept: () => {
-                this.products = this.products.filter(val => !this.selectedProducts.includes(val));
-                this.selectedProducts = null;
-                this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Products Deleted', life: 3000});
-            }
-        });
+        this.deleteProductsDialog = true;
     }
 
     editProduct(product: Product) {
@@ -81,16 +66,22 @@ export class AppCrudComponent implements OnInit {
     }
 
     deleteProduct(product: Product) {
-        this.confirmationService.confirm({
-            message: 'Are you sure you want to delete ' + product.name + '?',
-            header: 'Confirm',
-            icon: 'pi pi-exclamation-triangle',
-            accept: () => {
-                this.products = this.products.filter(val => val.id !== product.id);
-                this.product = {};
-                this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Deleted', life: 3000});
-            }
-        });
+        this.deleteProductDialog = true;
+        this.product = {...product};
+    }
+
+    confirmDeleteSelected(){
+        this.deleteProductsDialog = false;
+        this.products = this.products.filter(val => !this.selectedProducts.includes(val));
+        this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Products Deleted', life: 3000});
+        this.selectedProducts = null;
+    }
+
+    confirmDelete(){
+        this.deleteProductDialog = false;
+        this.products = this.products.filter(val => val.id !== this.product.id);
+        this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Deleted', life: 3000});
+        this.product = {};
     }
 
     hideDialog() {
@@ -103,12 +94,16 @@ export class AppCrudComponent implements OnInit {
 
         if (this.product.name.trim()) {
             if (this.product.id) {
+                // @ts-ignore
+                this.product.inventoryStatus = this.product.inventoryStatus.value ? this.product.inventoryStatus.value: this.product.inventoryStatus;
                 this.products[this.findIndexById(this.product.id)] = this.product;
                 this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Updated', life: 3000});
-            }
-            else {
+            } else {
                 this.product.id = this.createId();
+                this.product.code = this.createId();
                 this.product.image = 'product-placeholder.svg';
+                // @ts-ignore
+                this.product.inventoryStatus = this.product.inventoryStatus ? this.product.inventoryStatus.value : 'INSTOCK';
                 this.products.push(this.product);
                 this.messageService.add({severity: 'success', summary: 'Successful', detail: 'Product Created', life: 3000});
             }
@@ -134,7 +129,7 @@ export class AppCrudComponent implements OnInit {
     createId(): string {
         let id = '';
         const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        for ( let i = 0; i < 5; i++ ) {
+        for (let i = 0; i < 5; i++) {
             id += chars.charAt(Math.floor(Math.random() * chars.length));
         }
         return id;