+++ src/app/pages/app.help.component.ts
@@ -5,4 +5,16 @@ import { Component } from '@angular/core';
 })
 export class AppHelpComponent {
     text: any;
+
+    filteredText: any[];
+
+    filterText(event) {
+        const query = event.query;
+        this.filteredText = [];
+
+        for (let i = 0; i < 10; i++) {
+            this.filteredText.push(query + i);
+        }
+    }
+
 }