+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 13.1.0
+**Migration Guide**
+
+- Upgrade to PrimeFlex 3.1.2
+- Upgrade to Angular 13 and PrimeNG 13.1.0
 
 ## 12.2.0
 **Migration Guide**